Move original KeyringSessions to its own internal package
